Made the two Git runners extend the abstract parent
